# Make use of newlines index for performance gains