--- a/src/HaskellWorks/Data/Sv/Strict/Cursor.hs
+++ b/src/HaskellWorks/Data/Sv/Strict/Cursor.hs
@@ -1,5 +1,4 @@
 {-# LANGUAGE BangPatterns        #-}
-{-# LANGUAGE DeriveFunctor       #-}
 {-# LANGUAGE MultiWayIf          #-}
 {-# LANGUAGE ScopedTypeVariables #-}