Twice before:

```
$ time hw-sv query --source ~/arbornetwrks_v888_20180503.csv --create-index --column 0 --column 1 --delimiter , --target - > a.txt
149.06s user 45.39s system 99% cpu 3:15.98 total
```

Before:

```
$ time hw-sv query --source ~/arbornetwrks_v888_20180503.csv --create-index --column 0 --column 1 --delimiter , --target - > a.txt
147.67s user 47.63s system 91% cpu 3:34.41 total
```

The reason it is slower is because it is building an additional newlines index, which is currently not being used.

After:

```
$ time hw-sv query --source ~/arbornetwrks_v888_20180503.csv --create-index --column 0 --column 1 --delimiter , --target - --out-delimiter , > a.txt
374.33s user 51.32s system 272% cpu 2:36.47 total
```